requirements.txt
numpy==1.26.4
torch>=2.4
torchvision>=0.19
torchaudio>=2.4
ultralytics>=8.3.0
opencv-python
pillow
tkinterweb